Separate stories into features, add e2e tests.
